update: flexible examples on arbitrary input wasm
