+++ examples/v0/dialer/README.md
@@ -1,0 +1,17 @@
+# Dialer Example
+
+This is an example of using `water.Dialer` with any v0 WATM. 
+
+## Usage
+
+### Install
+
+```bash
+go get github.com/gaukas/water/examples/v0/dialer
+```
+
+### Run
+
+```bash
+dialer --raddr=<addr:port> --wasm=<path/to/watm.wasm>
+```
