+++ examples/v0/listener/README.md
@@ -1,0 +1,17 @@
+# Listener Example
+
+This is an example of using `water.Listener` with any v0 WATM. 
+
+## Usage
+
+### Install
+
+```bash
+go get github.com/gaukas/water/examples/v0/listener
+```
+
+### Run
+
+```bash
+listener --laddr=<addr:port> --wasm=<path/to/watm.wasm>
+```
