+++ examples/v0/listener/main.go
@@ -1,7 +1,9 @@
 package main
 
 import (
+	"context"
 	"crypto/rand"
+	"flag"
 	"fmt"
 	"net"
 	"os"
@@ -12,13 +14,15 @@ import (
 	_ "github.com/gaukas/water/transport/v0"
 )
 
-func main() {
-	if len(os.Args) != 2 {
-		panic("usage: dialer <remote_addr>")
-	}
-	var remoteAddr string = os.Args[1]
+var (
+	localAddr = flag.String("laddr", "", "local address to listen on")
+	wasmPath  = flag.String("wasm", "", "path to wasm file")
+)
 
-	wasm, err := os.ReadFile("./examples/v0/plain/plain.wasm")
+func main() {
+	flag.Parse()
+
+	wasm, err := os.ReadFile(*wasmPath)
 	if err != nil {
 		panic(fmt.Sprintf("failed to read wasm file: %v", err))
 	}
@@ -26,28 +30,33 @@ func main() {
 	// start using W.A.T.E.R. API below this line, have fun!
 	config := &water.Config{
 		TransportModuleBin: wasm,
-		NetworkDialerFunc:  net.Dial, // optional field, defaults to net.Dial
 	}
 	// configuring the standard out of the WebAssembly instance to inherit
 	// from the parent process
 	config.ModuleConfig().InheritStdout()
 	config.ModuleConfig().InheritStderr()
 
-	dialer, err := water.NewDialer(config)
+	lis, err := config.ListenContext(context.Background(), "tcp", *localAddr)
 	if err != nil {
-		panic(fmt.Sprintf("failed to create dialer: %v", err))
+		panic(fmt.Sprintf("failed to listen: %v", err))
 	}
-
-	conn, err := dialer.Dial("tcp", remoteAddr)
-	if err != nil {
-		panic(fmt.Sprintf("failed to dial: %v", err))
-	}
-	defer conn.Close()
-	// conn is a net.Conn that you are familiar with.
+	defer lis.Close()
+	log.Infof("Listening on %s:%s", lis.Addr().Network(), lis.Addr().String())
+	// lis is a net.Listener that you are familiar with.
 	// So effectively, W.A.T.E.R. API ends here and everything below
-	// this line is just how you treat a net.Conn.
+	// this line is just how you treat a net.Listener.
 
-	handleConn("server", conn)
+	clientCntr := 0
+	for {
+		conn, err := lis.Accept()
+		if err != nil {
+			panic(fmt.Sprintf("failed to accept: %v", err))
+		}
+
+		// start a goroutine to handle the connection
+		go handleConn(fmt.Sprintf("client#%d", clientCntr), conn)
+		clientCntr++
+	}
 }
 
 func handleConn(peer string, conn net.Conn) {
@@ -71,8 +80,8 @@ func handleConn(peer string, conn net.Conn) {
 		}
 	}()
 
-	// start a ticker for sending data every 2 seconds
-	ticker := time.NewTicker(2 * time.Second)
+	// start a ticker for sending message every 5 seconds
+	ticker := time.NewTicker(5 * time.Second)
 	defer ticker.Stop()
 
 	var sendBuf []byte = make([]byte, 4) // 4 bytes per message
@@ -80,6 +89,7 @@ func handleConn(peer string, conn net.Conn) {
 		select {
 		case msg := <-chanMsgRecv:
 			if msg == nil {
+				log.Warnf("read %s: connection closed, tearing down connection...", peer)
 				return // connection closed
 			}
 			log.Infof("%s: %x\n", peer, msg)
