--- examples/v0/plain/README.md
@@ -1,3 +1,0 @@
-# Plain
-
-This is just an example of using `water` with the `plain` transport, which does nothing to the bytestream but transmit it as-is.
