+++ examples/v0/relay/README.md
@@ -1,0 +1,17 @@
+# Relay Example
+
+This is an example of using `water.Relay` with any v0 WATM. 
+
+## Usage
+
+### Install
+
+```bash
+go get github.com/gaukas/water/examples/v0/relay
+```
+
+### Run
+
+```bash
+relay --laddr=<addr:port> --raddr=<addr:port> --wasm=<path/to/watm.wasm>
+```
