+++ examples/v0/watm/plain_rust.wasm
